refactor: merge the dotLottie feature set into lottie
