+++ examples/demo/Cargo.toml
@@ -8,7 +8,7 @@ publish = false
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
-bevy_vello = { path = "../../", features = ["experimental-dotLottie"] }
+bevy_vello = { path = "../../", features = ["lottie"] }
 bevy = { workspace = true }
 bevy_pancam = { version = "0.17.0", features = ["bevy_egui"] }
 bevy_egui = "0.32.0"
