--- src/integrations/dot_lottie/mod.rs
@@ -1,13 +1,0 @@
-mod lottie_player;
-pub use lottie_player::DotLottiePlayer;
-
-mod player_state;
-pub use player_state::PlayerState;
-
-mod player_transition;
-pub use player_transition::PlayerTransition;
-
-mod plugin;
-pub use plugin::DotLottieIntegrationPlugin;
-
-mod systems;
